refactor: #625 apply review feedback
See https://github.com/liferay/liferay-js-toolkit/pull/626#discussion_r469817860
