+++ packages/liferay-npm-bundler/src/steps/adapt/liferay-fragment/index.ts
